+++ chains/evm/deployment/v2_0_0/adapters/offramp_set_source_onramps_test.go
@@ -1,10 +1,9 @@
-package changesets
+package adapters
 
 import (
 	"testing"
 
 	"github.com/ethereum/go-ethereum/common"
-	cldf "github.com/smartcontractkit/chainlink-deployments-framework/deployment"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -36,22 +35,4 @@ func TestParseOffRampSourceOnRampAddresses_InvalidLength(t *testing.T) {
 
 	_, err := parseOffRampSourceOnRampAddresses([]string{"0x0102"})
 	require.Error(t, err)
-}
-
-func TestOffRampSetSourceOnRampsVerify(t *testing.T) {
-	t.Parallel()
-
-	verify := makeOffRampSetSourceOnRampsVerify()
-	env := cldf.Environment{}
-	require.Error(t, verify(env, OffRampSetSourceOnRampsInput{}))
-	require.Error(t, verify(env, OffRampSetSourceOnRampsInput{
-		Updates: []OffRampSetSourceOnRampsEntry{{LocalChainSelector: 1}},
-	}))
-	require.NoError(t, verify(env, OffRampSetSourceOnRampsInput{
-		Updates: []OffRampSetSourceOnRampsEntry{{
-			LocalChainSelector:  16015286601757825753,
-			SourceChainSelector: 10109143320554840099,
-			OnRamps:             []string{"0xa5f4d6b956c610c147282e1c180fcd04cfbed6cf8a0244289a1be44c7e784330"},
-		}},
-	}))
 }
